Add new umask-respecting functions to NamedTempFile
Prior to this commit, if you wanted to persist a NamedTempFile and have
it respect the current umask, you would need to set and reset the umask
in order to apply it to the persisted file (since umask(3p) returns
the umask, but requires you to provide a new one in order to get that
information).

This commit introduces new methods that allows the user the freedom of
respecting the umask at the time of the file's creation without having
to do a libc::umask dance.